requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pymongo>=4.6.0
motor>=3.3.2
dnspython>=2.4.2
schedule>=1.2.0
requests>=2.31.0
psutil>=5.9.6
cryptography>=42.0.0
python-socketio>=5.10.0
fastapi-socketio>=0.0.10
aiofiles>=23.2.1
email-validator>=2.1.0
jinja2>=3.1.2
PyJWT>=2.8.0
bcrypt>=4.1.2
aiohttp>=3.9.1
websockets>=12.0
